Added BLDTIME and Elapsed time. Major bugfix: combine terminals and intermediates to form targets. Cleanup for BuildAudit handling of new data Changed BuildAudit.update logic - now rewrites the DB key only in 'fresh' mode. Refactored GMakeCommand out to its own module
There are some builds which touch their prerequisites,
causing them to look like targets. To protect against
that we use the belt-and-suspenders approach of checking
against a list of files which predated the build.

Added --build-time option

Removed code which added dirs pointed to by symlinks to the dir list.
Should be unnecessary due to "pre-existing" check added in auditor module.

AuditMake.py -> AuditBuild.py